Style: 검색모달 토글 커스텀훅으로 만듦
